[rename] user.ts → User.ts
